feat:로그인페이지 배타그래프 완료. 상세페이지 배타그래프 미완.

+++ backend/routers/policy_router.py
@@ -90,8 +90,8 @@ def optimize_policies(
         db.query(Policy)
         .options(joinedload(Policy.tiers))
         .filter(Policy.is_active == True)
-        .filter((Policy.age_min == None) | (Policy.age_min.cast(Integer) <= age))
-        .filter((Policy.age_max == None) | (Policy.age_max.cast(Integer) >= age))
+        .filter((Policy.age_min == None) | (Policy.age_min <= age))
+        .filter((Policy.age_max == None) | (Policy.age_max >= age))
     )
 
     if request.profile.is_employed:
